requirement.txt
discord.py==2.6.0
yt-dlp
python-dotenv